chore: remove unused index entry point
