+++ index.html
@@ -14,6 +14,6 @@
         Press 'F' near car to Enter/Exit. <br>
         In Car: W (Accelerate), S (Brake/Reverse), A (Steer Left), D (Steer Right).
     </div>
-    <script type="module" src="index.tsx"></script>
+    <script type="module" src="./src/main.ts"></script>
 </body>
 </html>
